cross-images/entrypoint.sh: also bind-mount /tmp into Alpine chroot
As of recently, the rust link line includes a reference to some kind of
magical object file with a path of form `/tmp/rustcndJdVh/symbols.o`.
This didn't work for our Alpine chroot builds because `/tmp` started
meaning `/alpine/tmp`, which of course didn't contain the magic
temporary. We can easily work around this by bind-mounting `/tmp` into
the `/alpine` tree, in the same way that we do for many other
directories.